Allow updating existing records without a UUID
There is a case when you add `uuid_parameter` to an existing project: in
this case, you may have inconsistent values in the database. This commit
fixes handling such values and reflects specification to handle NULL
values and existing invalid values. It introduces `#reset_uuid!` which
is used in case of a missing UUID in the database.

Specifications are clearer and allowed to fix this bug where nil values
would supersede newly created values. Overall it's getting simpler.

For now it does not handle the case where a non-NULL value is present:
this can be the case when the database has been tampered with directly
without going through Rails. Ouch. In this case, the record cannot be
saved anymore (you can work around this using `#save(validate: false)`
or hitting the database with SQL directly.

However it's not handled automatically because it might be that your
existing UUIDs have been superseded, leaving your referential integrity
to gnaw cables at the bottom of the ocean (hat tip to /ETC in Bologna.)

This is going toward some refactoring with I18n integration and the
upcoming elimination of the validator (no need for so much code!).