Refactor: remove the validator class
And also...

- some refactoring to simplify code (removing the validator class)
- Bug fix: existing records with nil UUID would not get a new one
- clearer specs

New pre-release: v0.2.0-alpha.1

- a couple of issues to get the conversation started
  - on existing records with invalid UUIDs (#1)
  - on rails 5 plugins gems and I18n (#2)

Trying to figure out how to add locales into a Rails 5 plugin gem.
It seems to be related to I18n and Railtie ;o)

For now the translation files are loaded from the main .rb, but it's
certainly not the right way. The tests are failing on purpose to help
understand what's going on, but finding documentation that states this
simple thing resembles diving at the bottom of the ocean without light.